+++ demultiplex.h
@@ -25,6 +25,6 @@
 @protocol PPDecoderParent;
 @protocol Decode;
 
-int demultiplex_data(NSData *data, NSMutableArray *outlist, id <PPDecoderParent> parent, Class layer);
+size_t demultiplex_data(NSData *data, NSMutableArray *outlist, id <PPDecoderParent> parent, Class layer);
 
 #endif /* _DEMULTIPLEX_H_ */
